+++ client/client_test.go
@@ -218,10 +218,6 @@ func TestClient_Dial(t *testing.T) {
 					require.ErrorContains(t, err, "net/url: invalid control character in URL")
 					require.ErrorContains(t, err, "invalid server URI")
 				}},
-				{name: "missing port", s: "grpc://st1.storage.fs.neo.org", assert: func(t testing.TB, err error) {
-					require.ErrorContains(t, err, "missing port in address")
-					require.ErrorContains(t, err, "invalid server URI")
-				}},
 				{name: "invalid port", s: "grpc://st1.storage.fs.neo.org:foo", assert: func(t testing.TB, err error) {
 					require.ErrorContains(t, err, `invalid port ":foo" after host`)
 					require.ErrorContains(t, err, "invalid server URI")